+++ docs/make.jl
@@ -1,7 +1,29 @@
-using RedClust, Documenter
+using RedClust, Documenter, Literate
 
-DocMeta.setdocmeta!(RedClust, :DocTestSetup, :(using RedClust); recursive=true)
+# Parse the basic example file
+function preprocess_md(content)
+    content = replace(content, "#setup!" => "# ```@setup @__NAME__\n#=") # hacky workaround to create @setup blocks in Literate
+    content = replace(content, "#!setup" => "=#\n# ```")
+    return content
+end
+function preprocess_jl(content)
+    content = replace(content, "#setup!" => "")
+    content = replace(content, "#!setup" => "")
+    content = replace(content, r" *nothing *# *hide *(\r\n|\n)" => "") # remove the nothing # hide lines when writing to .jl files
+    return content
+end
+inputdir = joinpath(@__DIR__, "..", "examples")
+inputfile = joinpath(inputdir, "_basic_example.jl")
+outputdir = joinpath(@__DIR__, "src", "_generated")
+# Create the example file in the docs
+Literate.markdown(inputfile, outputdir; 
+name = "basic_example", preprocess = preprocess_md)
+# Create the actual example file
+Literate.script(inputfile, outputdir; 
+name = "basic_example", preprocess = preprocess_jl, keep_comments=true)
 
+# Build the documentation HTML pages
+Documenter.Writers.HTMLWriter.HTML(ansicolor = true)
 makedocs(;
     modules=[RedClust],
     authors="Abhinav Natarajan <abhinav.v.natarajan@gmail.com>",
@@ -12,9 +34,11 @@ makedocs(;
         canonical="https://abhinavnatarajan.github.io/RedClust.jl",
         edit_link="main",
         assets=String[],
+        ansicolor=true
     ),
     pages=[
     "Introduction" => "index.md",
+    "Example" => joinpath("_generated", "basic_example.md"),
     "Reference" => "reference.md",
     "Changelog" => "changelog.md"
     ]