+++ examples/basic_example.jl
@@ -1,26 +1,116 @@
+#md # # Example
+# Here we demonstrate the usage of RedClust through an example. 
+#md # This example can be downloaded as a Julia script [here](./@__NAME__.jl).
+# We begin by setting up the necessary includes. 
+
 using RedClust, Plots, StatsPlots
 using Random: seed!
 using StatsBase: counts
-include("utils_for_examples.jl")
+using LinearAlgebra: triu, diagind
 
-########## Generate data ##########
+#jl ## We define some colors optimized for color-blind individuals based on [this article](https://www.nature.com/articles/nmeth.1618) by Bang Wong in Nature. 
+#setup!
+function wong_colors(alpha = 1.0)
+    colors = [
+        RGB(0/255, 114/255, 178/255), # blue
+        RGB(230/255, 159/255, 0/255), # orange
+        RGB(0/255, 158/255, 115/255), # green
+        RGB(204/255, 121/255, 167/255), # reddish purple
+        RGB(86/255, 180/255, 233/255), # sky blue
+        RGB(213/255, 94/255, 0/255), # vermillion
+        RGB(240/255, 228/255, 66/255), # yellow
+    ]
+    @. RGBA{Float32}(red(colors), green(colors), blue(colors), alpha)
+end
+## Set plotting defaults
+default(fontfamily = "Computer Modern",
+color_palette = wong_colors(0.8), 
+gridlinewidth = 1,
+framestyle = :box,
+linecolor = :match,
+linewidth = 0.5,
+guidefontsize = 16, 
+tickfontsize = 16, 
+colorbar_tickfontsize = 16, 
+legend_font_pointsize = 16)
+#!setup
+
+# Next we define some convenience functions for plotting.
+## Heatmap of square matrix
+function sqmatrixplot(X::Matrix; kwargs...)
+    M, N = size(X)
+    heatmap(
+        X, 
+        aspect_ratio=:equal, 
+        color=:Blues, 
+        xlim=(1,M), ylim=(1,N), 
+        yflip = true, xmirror=true;
+        kwargs...)
+end
+
+## Histogram with integer bins
+function histogram_pmf(X::AbstractVector{<:Integer}; kwargs...)
+    xvals = minimum(X):maximum(X)
+    yvals = counts(X)./length(X)
+    bar(xvals, yvals, 
+    linewidth = 0, 
+    legend = false, 
+    xticks = xvals; kwargs...)
+end
+
+## Combine two symmetric square matrices together into the upper and lower triangle of a square matrix
+function combine_sqmatrices(lower::Matrix, upper::Matrix, diagonal::String = "lower") 
+    if size(lower)[1] != size(lower)[2]
+        throw(ArgumentError("Argument `lower` must be square, has dimensions $(size(lower))."))
+    end
+    if size(upper)[1] != size(upper)[2]
+        throw(ArgumentError("Argument `upper` must be a square matrix, has dimensions $(size(upper))."))
+    end
+    if !all(size(lower) .== size(upper))
+        throw(ArgumentError("Arguments `lower` and `upper` must have the same size."))
+    end
+    if !(eltype(lower) <: eltype(upper)) && !(eltype(upper) <: eltype(lower))
+        throw(ArgumentError("Arguments must have compatible entries, got $(eltype(lower)) and $(eltype(upper))."))
+    end
+    if diagonal ∉ ["lower", "upper"]
+        throw(ArgumentError("Keyword argument `diagonal` must be either \"lower\" or \"upper\"."))
+    end
+    result = copy(lower)
+    temp = trues(size(lower))
+    upper_idx = triu(temp, 1)
+    diagonal_idx = diagind(temp)
+    result[upper_idx] .= upper[upper_idx]
+    result[diagonal_idx] .= ((diagonal == "lower") ? lower : upper)[diagonal_idx]
+    return result
+end
+nothing # hide
+
+# ## Generating Data
+# We can generate some example data using the function [`generatemixture`](@ref). 
 begin
     K = 10 # Number of clusters 
     N = 100 # Number of points
     data_σ = 0.25 # Variance of the normal kernel
     data_dim = 10 # Data dimension
     α = 10 # parameter for Dirichlet prior on cluster weights
-    rng = 44
     data = generatemixture(N, K; 
-    α = α, σ = data_σ, dim = data_dim, rng)
+    α = α, σ = data_σ, dim = data_dim)
     points, distmatrix, clusts, probs, oracle_coclustering = data
+    nothing # hide 
 end
-# Plot the true adjacency matrix and oracle co-clustering matrices as heatmaps
+# Alternatively, the function [`example_dataset`](@ref) can be used to retrieve the datasets used in the original RedClust paper.  
+begin
+    data = example_dataset(1)
+    points, distmatrix, clusts, probs, oracle_coclustering = data
+    nothing # hide
+end
+# We can visualise the true adjacency matrix of the observations with respect to the true clusters that they were drawn from.
 sqmatrixplot(adjacencymatrix(clusts), title = "Adjacency Matrix")
+# We can visualise the oracle co-clustering matrix. This matrix is the matrix of co-clustering probabilities of the observations conditioned upon the data generation process. This takes into account full information about the cluster weights (and how they are generated), the mixture kernels for each cluster, and the location and scale parameters for these kernels. 
 sqmatrixplot(oracle_coclustering, title = "Oracle Coclustering Probabilities")
-# Visualise distance matrix
+# We can visualise the distance matrix of the observations.
 sqmatrixplot(distmatrix)
-# Plot histogram of distances
+# We can also plot the histogram of distances, grouped by whether they are inter-cluster distances (ICD) or within-cluster distances (WCD).
 begin 
     empirical_intracluster = uppertriangle(distmatrix)[
         uppertriangle(adjacencymatrix(clusts)) .== 1]
@@ -34,14 +124,13 @@ begin
     bins = minimum(empirical_intracluster):0.05:maximum(empirical_intracluster),
     label="WCD")
 end
-
-############### Prior fitting ##############
-# Determine the best the prior hyperparameters
+#md # ## Prior Hyperparameters
+# RedClust includes the function [`fitprior`](@ref) to heuristically choose prior hyperparameters based on the data.
 params = fitprior(points, "k-means", false)
-# Plot Empirical vs prior predictive density of distances
+# We can check how good the chosen prior hyperparameters are by comparing the empirical distribution of distances to the predictive distribution based on the prior. 
 begin 
-    pred_intracluster = sampledist(params, 10000, "intracluster")
-    pred_intercluster = sampledist(params, 10000, "intercluster")
+    pred_intracluster = sampledist(params, "intracluster", 10000)
+    pred_intercluster = sampledist(params, "intercluster", 10000)
     density(pred_intracluster, 
     label="Simulated WCD", xlabel = "Distance", ylabel = "Density", 
     size = (700, 500), 
@@ -56,40 +145,28 @@ begin
     label="Empirical ICD", 
     linewidth = 2, primary = false)
 end
-# Visualise the marginal distribution on K
+# We can also evaluate the prior hyperparameters by checking the marginal predictive distribution on ``K`` (the number of clusters). 
 begin
     Ksamples = sampleK(params, 10000, N)
-    density(Ksamples, xlabel = "K", ylabel = "Density", linewidth = 2, legend = false, title = "Marginal Prior Predictive Density of K")
+    density(Ksamples, linewidth = 2, legend = false, 
+    xlabel = "K", ylabel = "Density", title = "Marginal Prior Predictive Density of K")
 end
+#md # ## Sampling
+# Running the MCMC is straightforward. We set up the MCMC options using [`MCMCOptionsList`](@ref).
+options = MCMCOptionsList(numiters=5000)
+# We then set up the input data using [`MCMCData`](@ref).
+data = MCMCData(points)
+# We can then run the sampler using [`runsampler`](@ref). 
+result = runsampler(data, options, params)
 
-
-############### MCMC ##############
-# MCMC options
-begin
-    options = MCMCOptionsList(
-        numiters=5000)
-    data = MCMCData(points)
-end
-
-# Run the sampler
-begin
-    result = runsampler(data, options, params)
-    pointestimate, index = getpointestimate(result; loss = "binder", method="MAP")
-end
-
-######## Check the results ########
-# Summary of point estimate
-summarise(pointestimate, clusts)
-# Posterior coclustering matrix
+#md # ## MCMC Result
+# The MCMC result contains several details about the MCMC, including acceptance rate, runtime, and convergence diagnostics. For full details see [`MCMCResult`](@ref). In this example we have the ground truth cluster labels, so we can evaluate the result. For example, we can compare the posterior coclustering matrix to the oracle co-clustering probabilities.
 sqmatrixplot(combine_sqmatrices(result.posterior_coclustering, oracle_coclustering), 
 title="Posterior vs Oracle Coclustering Probabilities")
-# Point-estimate adjacency matrix
-sqmatrixplot(combine_sqmatrices(adjacencymatrix(pointestimate), adjacencymatrix(clusts)), 
-title = "True Clustering vs MAP Point Estimate")
-# Posterior distribution of K
+# Plot the posterior distribution of K:
 histogram_pmf(result.K, xlabel = "K", ylabel = "PMF", 
 size = (400, 400), title = "Posterior Distribution of K")
-# Posterior distribution of r
+# Plot the posterior distribution of r:
 begin
     histogram(result.r, normalize = :pdf,
     legend_font_pointsize=12, 
@@ -99,7 +176,7 @@ begin
     color=:black, linewidth = 2, linestyle=:dash, 
     label="Kernel estimate", legend_font_pointsize=12)
 end
-# Posterior distribution of p
+# Plot the posterior distribution of p:
 begin
     histogram(result.p, normalize = :pdf, 
     ylabel = "Density", xlabel = "p", 
@@ -108,11 +185,19 @@ begin
     density!(result.p, color=:black, linewidth = 2, linestyle=:dash, 
     label = "Kernel estimate")
 end
-# Log-likelihood
+# Check the trace plot of the log-likelihood to make sure the MCMC is moving well:
 plot(result.loglik, legend = false, linewidth = 1,
 xlabel = "Iteration", ylabel = "Log likelihood", 
 title = "Log-Likelihood Trace Plot")
-# Log-posterior
+# Check the trace plot of the log-posterior:
 plot(result.logposterior, legend = false, linewidth = 1,
 xlabel = "Iteration", ylabel = "Log posterior", 
 title = "Log-Posterior Trace Plot")
+#md # ## Point Estimates
+# The function [`getpointestimate`](@ref) finds an optimal point estimate, based on some notion of optimality. For example, to get the maximum a posteriori estimate we can run the following.
+pointestimate, index = getpointestimate(result; method="MAP")
+# We can compare the point-estimate to the true clustering through their adjacency matrices.
+sqmatrixplot(combine_sqmatrices(adjacencymatrix(pointestimate), adjacencymatrix(clusts)), 
+title = "True Clustering vs MAP Point Estimate")
+# We can check the accuracy of the point estimate in terms of clustering metrics.
+summarise(pointestimate, clusts)